handler added and consumer binary added
